# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion RH API — ARTF",
        "description": "API REST de gestion RH (intégration administrative des agents).\n\n## Authentification\n\n1. Appeler `POST /api/login` avec email / mot de passe\n2. Copier le `token` retourné\n3. Cliquer **Authorize** en haut à droite\n4. Saisir : `Bearer {token}` (ou uniquement le token selon le client)\n\nComptes de test (seeders) :\n- `admin@arft.cg` / `Admin@2026`\n- `rh@arft.cg` / `Rh@2026`\n\n## Convention de réponse\n\n```json\n{ \"data\": {}, \"message\": \"...\" }\n```\n\nErreurs : `{ \"success\": false, \"message\": \"...\", \"errors\": {} }`",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API (variable L5_SWAGGER_CONST_HOST)"
        }
    ],
    "paths": {
...
        "/api/integration/nominations": {
            "get": {
                "tags": [
                    "Intégration — Nominations"
                ],
                "summary": "Liste des nominations",
                "operationId": "listNominations",
                "responses": {
                    "200": {
                        "description": "Liste"
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error401"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Intégration — Nominations"
                ],
                "summary": "Créer une nomination",
                "operationId": "storeNomination",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "agent_id",
                                    "poste",
                                    "structurable_type",
                                    "structurable_id",
                                    "date_debut"
                                ],
                                "properties": {
                                    "agent_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "poste": {
                                        "type": "string",
                                        "example": "Chef de Service",
                                        "enum": [
                                            "Directeur Général",
                                            "Directeur Central",
                                            "Directeur Départemental",
                                            "Chef de Service",
                                            "Chef de Bureau"
                                        ]
                                    },
                                    "structurable_type": {
                                        "type": "string",
                                        "example": "App\\Models\\Service"
                                    },
                                    "structurable_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date_debut": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-09-01"
                                    },
                                    "type_acte": {
                                        "type": "string",
                                        "nullable": true,
                                        "enum": [
                                            "arrete",
                                            "decision",
                                            "note_service"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NominationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error422"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/integration/nominations/{id}": {
            "get": {
                "tags": [
                    "Intégration — Nominations"
                ],
                "summary": "Détail d'une nomination",
                "operationId": "showNomination",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NominationResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error404"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/integration/nominations/{nomination}/activer": {
            "post": {
                "tags": [
                    "Intégration — Nominations"
                ],
                "summary": "Activer une nomination",
                "operationId": "activerNomination",
                "parameters": [
                    {
                        "name": "nomination",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "dossier_integration_id": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Activée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NominationResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/integration/nominations/{nomination}/cloturer": {
            "post": {
                "tags": [
                    "Intégration — Nominations"
                ],
                "summary": "Clôturer une nomination",
                "operationId": "cloturerNomination",
                "parameters": [
                    {
                        "name": "nomination",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "date_fin": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Clôturée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NominationResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/integration/nominations/{nomination}/rejeter": {
            "post": {
                "tags": [
                    "Intégration — Nominations"
                ],
                "summary": "Rejeter une nomination",
                "operationId": "rejeterNomination",
                "parameters": [
                    {
                        "name": "nomination",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransitionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rejetée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NominationResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/integration/agents/{agent}/nominations": {
            "get": {
                "tags": [
                    "Intégration — Nominations"
                ],
                "summary": "Nominations d'un agent",
                "operationId": "listNominationsByAgent",
                "parameters": [
                    {
                        "name": "agent",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Error401": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Non authentifié. Veuillez vous connecter."
                    }
                },
                "type": "object"
            },
...
            "Error404": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Ressource introuvable."
                    }
                },
                "type": "object"
            },
            "Error422": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Les données fournies sont invalides."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "Le champ email est obligatoire."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "Nomination": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "agent_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "poste": {
                        "type": "string",
                        "example": "Chef de Service"
                    },
                    "structurable_type": {
                        "type": "string",
                        "example": "App\\Models\\Service"
                    },
                    "structurable_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "date_debut": {
                        "type": "string",
                        "format": "date"
                    },
                    "date_fin": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "type_acte": {
                        "type": "string",
                        "example": "decision",
                        "nullable": true
                    },
                    "statut": {
                        "type": "string",
                        "example": "EN_ATTENTE"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "NominationResponse": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Nomination"
                    },
                    "message": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "TransitionRequest": {
                "properties": {
                    "commentaire": {
                        "type": "string",
                        "example": "Pièces manquantes : CV et diplôme",
                        "nullable": true,
                        "maxLength": 1000
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token Sanctum obtenu via POST /api/login",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}